refactor: 同じ関数を共通化
