+++ app/routes/admin.timeline.tsx
@@ -3,6 +3,7 @@ import { Link, redirect, useLoaderData } from "@remix-run/react";
 import { useState } from "react";
 import { prismaClient } from "~/.server/prisma";
 import { getAdmin } from "~/.server/supabase";
+import { getMonths } from "~/utils/date";
 import { dayjsJP } from "~/utils/dayjs";
 
 export const meta = () => [
@@ -43,22 +44,6 @@ export const loader = async ({ context, request }: LoaderFunctionArgs) => {
 		},
 	});
 	return { purchases: purchases.reverse() };
-};
-
-const getMonths = (oldest: Date, newest: Date) => {
-	const dayjs = dayjsJP();
-	const oldestDate = dayjs(oldest).tz().startOf("month");
-	const newestDateNextMonth = dayjs(newest)
-		.tz()
-		.startOf("month")
-		.add(1, "month");
-	const months = [];
-	let current = oldestDate;
-	while (current.isBefore(newestDateNextMonth)) {
-		months.push(current);
-		current = current.add(1, "month");
-	}
-	return months;
 };
 
 export default function Timeline() {
